Move BedParser to seperate file
