+++ scripts/BamToMutposVCF.py
@@ -4,7 +4,7 @@ import sys
 from argparse import ArgumentParser
 from collections import Counter
 import pysam
-
+from BedParser import *
 
 def SamHeaderToVcfHeader(inSamHeader, progName, progVersion, progCmd):
     headLines = [
@@ -42,77 +42,6 @@ def SamHeaderToVcfHeader(inSamHeader, progName, progVersion, progCmd):
     headLines.append(f"##{progName}Command={progCmd}")
     headLines.extend(contigBlock)
     return headLines
-
-
-class Bed_File:
-    def __init__(self, inFile):
-        self.file = open(inFile, 'r')
-
-    def __iter__(self):
-        return self
-
-    def __next__(self):
-        myLine = self.file.readline()
-        if myLine != "":
-            return Bed_Line(*myLine.strip().split())
-        else:
-            raise StopIteration
-
-
-class Bed_Line:
-    def __init__(self,
-                 chrom,
-                 startPos,
-                 endPos,
-                 name="",
-                 score="",
-                 strand=".",
-                 thickStart="",
-                 thickEnd="",
-                 itemRGB="",
-                 blockCount="",
-                 blockSizes="",
-                 blockStarts=""
-                 ):
-        self.chrom = chrom
-        self.startPos = int(startPos)
-        self.endPos = int(endPos)
-        self.name = name
-        if score == '':
-            self.score = 0
-        else:
-            self.score = int(score)
-        if strand in (".", "+", "-"):
-            self.strand = strand
-        else:
-            self.strand = "."
-        if thickStart == "":
-            self.thickStart = self.startPos
-        else:
-            self.thickStart = int(thickStart)
-        if thickEnd == "":
-            self.thickEnd = self.endPos
-        else:
-            self.thickEnd = int(thickEnd)
-        if itemRGB == "":
-            self.itemRBG = None
-        else:
-            self.itemRGB = itemRGB
-        if blockCount == "":
-            self.blockCount = 1
-        else:
-            self.blockCount = int(blockCount)
-        if blockSizes == "":
-            self.blockSizes = [abs(self.endPos - self.startPos)]
-        else:
-            self.blockSizes = [int(x) for x in blockSizes.split(',')]
-        if blockStarts == "":
-            self.blockStarts = 0
-        else:
-            self.blockStarts = [int(x) for x in blockStarts.split(',')]
-
-    def samtoolsStr(self):
-        return f"{self.chrom}:{self.startPos}:{self.endPos}"
 
 
 def getParams():
